Removed mouse move event forwarding to Swing UI layer in Windows
This change was implemented as part of POS-2897: Optimize (debugging) performance of nxt-chromium, particularly on Windows. It exempts mouse move events from being forwarded to the CefWindowHandler implementation entirely, because there appears to be no actual reason to have these events forwarded, at least not in our particular use case. One key insight justifying this change has been that apparently, the particular forwarding of mouse events to CefWindowHandler is only implemented on Windows and not supported on MacOS and Linux at all, of which at least Linux is a very important target OS for us, which means that any feature we rely on would have to work on Linux as well.